+++ Events.php
@@ -4,6 +4,7 @@ namespace humhub\modules\darkMode;
 
 use humhub\modules\darkMode\widgets\DarkStyle;
 use humhub\modules\darkMode\widgets\SwitchButton;
+use humhub\components\ModuleEvent;
 use Yii;
 
 class Events
@@ -21,4 +22,25 @@ class Events
             Yii::error($e);
         }
     }
+    
+    public static function onAfterModuleEnabled(ModuleEvent $event)
+    {
+        // If module ID contains "theme" we assume it is a theme module
+        if(strpos($event->moduleId, 'theme') !== false) {
+            // Delete module setting to prevent design issues
+            $settings = Yii::$app->getModule('dark-mode')->settings;
+            $settings->delete('theme');
+        }
+    }
+
+    public static function onBeforeModuleDisabled(ModuleEvent $event)
+    {
+        $settings = Yii::$app->getModule('dark-mode')->settings;
+        $darkTheme = $settings->get('theme');
+        
+        // If Enterprise was disabled, remove "DarkEnterprise" setting
+        if ($event->moduleId == 'enterprise-theme' && $darkTheme == 'DarkEnterprise') {
+            $settings->delete('theme');
+        }
+    }    
 }