+++ assets/DarkStyleAsset.php
@@ -3,20 +3,19 @@
 namespace humhub\modules\darkMode\assets;
 
 use humhub\modules\darkMode\models\Config;
-use humhub\modules\ui\view\helpers\ThemeHelper;
 use yii\web\AssetBundle;
 
 /*
  * DarkStyleAsset is the default dark style, based on browser/system preference
+ * It should only be registered within try and catch because it throws an error when no theme is found
  */
-
 class DarkStyleAsset extends AssetBundle
 {
     public $publishOptions = [
         'forceCopy' => false
     ];
 
-    public $sourcePath = '@dark-mode/themes/DarkHumHub';
+    public $sourcePath;
 
     // Tell browser to use stylesheet only when in dark mode
     public $css = [
@@ -25,10 +24,9 @@ class DarkStyleAsset extends AssetBundle
 
     public function init()
     {
+        parent::init();
         // Find theme selected by module settings
         $config = new Config();
-        $theme = ThemeHelper::getThemeByName($config->theme);
-
-        $this->sourcePath = $theme->basePath;
+        $this->sourcePath = $config->getThemePath();
     }
 }