+++ widgets/DarkStyle.php
@@ -6,6 +6,7 @@ use humhub\components\Widget;
 use humhub\modules\darkMode\assets\DarkStyleAsset;
 use humhub\modules\darkMode\assets\ForceDarkStyleAsset;
 use humhub\modules\darkMode\models\UserSetting;
+use Yii;
 
 /**
  * Adds the dark style asset
@@ -18,11 +19,17 @@ class DarkStyle extends Widget
 
         $userSettings = new UserSetting();
 
-        // Register the right asset
-        if ($userSettings->darkMode === UserSetting::OPTION_DEFAULT) {
-            DarkStyleAsset::register($view);
-        } elseif ($userSettings->darkMode === UserSetting::OPTION_DARK) {
-            ForceDarkStyleAsset::register($view);
+        // Try to register the right asset
+        try {
+            if ($userSettings->darkMode === UserSetting::OPTION_DEFAULT) {
+                DarkStyleAsset::register($view);
+            } elseif ($userSettings->darkMode === UserSetting::OPTION_DARK) {
+                ForceDarkStyleAsset::register($view);
+            }
+        } catch (\Throwable $e) {
+            Yii::error('Stylesheet for Dark Mode could not be loaded. Please check the module configuration! You probably have to select a dark theme. 
+                Full error:
+                ' . $e, 'dark-theme');
         }
 
         return '';