+++ widgets/views/switchButton.php
@@ -1,11 +1,8 @@
 <?php
 
-use humhub\modules\darkMode\assets\DarkStyleAsset;
-use yii\helpers\Url;
-
 ?>
 <div class="btn-group">
-    <a id="dark-mode-button" href="#" data-action-click="ui.modal.load" data-action-click-url="<?= Url::toRoute('/dark-mode/user/') ?>">
+    <a id="dark-mode-button" href="#" data-action-click="ui.modal.load" data-action-click-url="<?= $url ?>">
         <?= $icon ?>
     </a>
 </div>