+++ docs/CHANGELOG.md
@@ -1,5 +1,12 @@
 Changelog
 =========
+## 1.0.0-beta3 (4/8/2023)
+- Enh: Compatibility with Enterprise Theme
+- allow empty module setting & log error if asset cannot be registered
+- HumHub minVersion: 1.15
+- DarkHumHub: CSS compiled with HumHubb 1.15
+- DarkHumHub: updated color text-color-soft3
+
 ## 1.0.0-beta2 (25/7/2023)
 - [#7](https://github.com/felixhahnweilheim/humhub-dark-mode/pull/7) Replaced cookies with session and small fixes
 - Additionally save setting as user setting to database (if logged in)